add controllers folder to classify sortArray and admin logic
